Base sleep sound display limit on num sounds available
Rather than using a hard limit (i.e. 5 or 11) to display sounds, ensure that we don't return sounds
to the user unless their Sense has all of the files it is _supposed_ to have.

This has the added benefit of allowing us to "remove" files server-side without deleting the row in
the DB or deleting the file from Sense.